+++ jgclark.DashboardReact/src/react/components/CommandButton.jsx
@@ -7,16 +7,17 @@
 import React from 'react'
 import type { TActionButton } from '../../types.js'
 import { useAppContext } from './AppContext.jsx'
-import {logDebug,JSP} from '@helpers/react/reactDev.js'
+import { logDebug, JSP } from '@helpers/react/reactDev.js'
 
 type ButtonProps = {
   button: TActionButton,
+  onClick: (button: TActionButton) => void, // send this button info back up
   // param: string,
 }
 
 function CommandButton(inputObj: ButtonProps): React$Node {
   const { sendActionToPlugin /*, sendToPlugin, dispatch, pluginData */ } = useAppContext()
-  const { button } = inputObj
+  const { button, onClick } = inputObj
 
   // logDebug(`CommandButton`,`setting up button: ${button.display}, button=${JSP(button,2)}`)
 
@@ -25,17 +26,23 @@ function CommandButton(inputObj: ButtonProps): React$Node {
   // const possIconAfter = (button.iconAfter !== '') ? <i className={`padLeft ${button.iconAfter}`}></i> : ''
   // Instead will use dangerouslySetInnerHTML, so we can set anything.
 
+  const handleButtonClick = () => {
+    logDebug(`CommandButton`, `button clicked: ${button.display}`)
+    sendActionToPlugin(button.actionPluginID, {
+      actionType: button.actionName,
+      toFilename: button.actionParam,
+      sectionCodes: button.postActionRefresh
+    })
+    onClick(button)
+  }
+
   return (
     <>
       {' '}
       <button
         className="PCButton tooltip"
         data-tooltip={button.tooltip}
-        onClick={() => sendActionToPlugin(button.actionPluginID, {
-          actionType: button.actionName,
-          toFilename: button.actionParam,
-          sectionCodes: button.postActionRefresh
-        })}
+        onClick={handleButtonClick}
         dangerouslySetInnerHTML={{ __html: button.display }}
       >
       </button>