+++ np.Shared/src/react/Root.jsx
@@ -311,7 +311,7 @@ export function Root(/* props: Props */): Node {
       if (!action) throw new Error('sendToPlugin: command/action must be called with a string')
       // logDebug(`Root`, ` sendToPlugin: ${JSON.stringify(action)} ${additionalDetails}`, action, data, additionalDetails)
       if (!data) throw new Error('sendToPlugin: data must be called with an object')
-      console.log(`Root`, ` sendToPlugin: command:${action} data=${JSON.stringify(data)} `)
+      // logDebug(`Root`, ` sendToPlugin: command:${action} data=${JSON.stringify(data)} `)
       const { command, id } = returnPluginCommand // this comes from the initial data passed to the plugin
       runPluginCommand(command, id, [action, data, additionalDetails])
     },